// fe/src/lib/api/generated/model/processingActivityUpdateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface ProcessingActivityUpdateRequest {
  authorized_access_summary?: string | null;
  data_subject_categories?: string[] | null;
  name?: string | null;
  personal_data_categories?: string[] | null;
  policy_basis_reference?: string | null;
  purpose?: string | null;
  retention_policy_reference?: string | null;
  safeguards_summary?: string | null;
}